Campus-Recruitement-Project
The Campus Recruitment System is a minor project developed by Rahul Patel and Rishi
Raghuwansh from Oriental Institute of Science and Technology in Bhopal. This project aims to
streamline the campus recruitment process by providing an efficient system for both employers and
students. The abstract of this project will provide a concise overview of its objectives, features, and
expected outcomes.
The system will include features such as employer registration, student registration, job posting,
application submission, and interview scheduling. It will also generate reports and statistics to track
the progress and effectiveness of the recruitment process. The Campus Recruitment System will
significantly reduce the manual effort involved in organizing and managing campus recruitment
activities, making it more streamlined and convenient for all stakeholders involved.
The system will also have a user-friendly interface that will allow employers and students to easily
navigate and interact with the platform. Additionally, it will incorporate a notification system to
keep both employers and students updated on the status of their applications and interviews. With
the implementation of the Campus Recruitment System, the recruitment process will become more
efficient and transparent, facilitating better communication and decision-making.
Moreover, the system will have a robust search functionality to enable employers to find suitable
candidates based on specific criteria such as education, skills, and experience. It will also provide
personalized recommendations to students, suggesting relevant job opportunities based on their
profiles and interests. By incorporating these features, the Campus Recruitment System aims to
enhance the overall experience for both employers and students, ensuring better matches and
increased success rates in the recruitment process.
The Placement Cell maintains a large database of students wherein all the information of students
including personal records and the academic performance in terms C.P.I. is stored and company
information including profile of company, eligibility criteria and facilities it provides etc. The
software retrieves this data and displays as per the user requirement.
The Placement Management System is developed as an attempt to take a record of company and
students by restricting such a large database to that of a particular class of students or company. The
system provides the facility of viewing both the personal and academic information of the student
and company. Company can post vacancy and students can apply if they are eligible. Company can
view the applications and can accept or reject them.
The project is to facilitate students in college, company to register and communicate with Placement
Cell.